% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.R
\name{alpha_div}
\alias{alpha_div}
\title{Alpha diversity measure}
\usage{
alpha_div(physeq, method)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{method}{(Required). A list of character strings specifying \code{method} to be used to calculate for alpha diversity
in the data. Available methods are: "richness", "fisher", "simpson", "shannon" and "evenness".}
}
\value{
It returns a data frame of diversity measure and corresponding indices/methods
}
\description{
This function calculates alpha diversity of provided community data using
selected indices/method.
}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
