% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurence_network.R
\name{co_occurence_network}
\alias{co_occurence_network}
\title{Co-occurence network}
\usage{
co_occurence_network(physeq, qval_threshold = 0.05, grouping_column,
  rhos = c(-0.75, -0.5, 0.5, 0.75), select.condition = NULL,
  method = "cor", filename = NULL, ...)
}
\arguments{
\item{qval_threshold}{Cut off for "fdr" adjusted p-values.}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the information.
information.}

\item{rhos}{(required). A vector specifying thresholds for correlation between co-occuring pairs of features.}

\item{select.condition}{(optional). A character string speifying name of a desired condition/group. If not supplied, co-occuence analysis
is performed amongst all conditions present in the grouping variable.}

\item{physeq(Required).}{A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{method.}{A character string for the correlation method to be used; options include: "cor" and "bicor"}
}
\value{
Files of visual representation of the network showing subcommunities (identified by colors),
network statistics and file containing pairwise corrrelations of taxa in under different conditions.
}
\description{
This function uses co-occurence pattern analysis to identify co-occuring features/taxa in community
data under specified environmental conditions. Co-occurence is measured as positive correlation whose threshold(s)
can be specified as indicated in arguments section. Amongst these features, pairwise co-occurences which are outstanding within sub communities
are detected. p-values generated during pairwise correlations are adjusted for multiple comparisons
by false discovery rate. The network statistics used to assign importance of taxa/features include betweenness, closeness,
and eigenvector centrality.
}
\examples{
data(pitlatrine)
physeq <- taxa_level(pitlatrine, "Phylum")
co_occr <- co_occurence_network(physeq, grouping_column = "Country", rhos = 0.35, select.condition = "V", scale.vertex.size=3, scale.edge.width=15)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015

[Ryan J. Williams, Adina Howe and Kirsten S. Hofmockel.
           Demonstrating microbial co-occurrence pattern analyses within and between ecosystems,
           Frontiers in Microbial Ecology, 5:358, 2014].
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
