% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{differential_abundance}
\alias{differential_abundance}
\title{Differential abundance analysis}
\usage{
differential_abundance(physeq, grouping_column, pvalue.threshold = 0.05,
  lfc.threshold = 0, filename = "NB_significant", output_norm = NULL)
}
\arguments{
\item{physeq}{(Required). A link[phyloseq]{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the.
information, this should be one of the components of grouping vector.}

\item{output_norm}{(optional). A character string specifying method to be used for transforming abundance data to be used for plotting.
note that, this normalisation occurs after DESeq analysis. Therefore, it is strictly for purposes of the output data
and not for differential expression analysis.}

\item{pvalue.threshold.}{Cut off p-value for significance of differentially abundance taxa, default is 0.05.}

\item{lfc.threshold.}{Threshold for log2 fold change over which significance of differentially expressed taxa is considered.}
}
\value{
Returns a ggplot object. This can further be manipulated as preferred by user.

Returns a list of three items: \itemize{
        \item  SignfeaturesTable: A \code{data.frame} of taxa with raw and adjusted pvalues, basemean, log2 fold change and significance labels.
        computed by using raw p-values.
        \item  importance: A \code{data.frame} of taxa mean decrease in accuracy as obtained by random forest classifier.
        \item plotdata: A \code{data.frame} of taxa and corresponding corrected p-values, importance rank organised
        in form accepted by ggplot.
       }
}
\description{
This function finds the features that are significantly differentially abundant in the provided data,
using DESeq implementation which models taxa abundance as a negative binomial distribution. See \link[DESeq2]{DESeq}  for more details. The significance
of differentially abundant taxa is defined by log2 fold change and pvalue thresholds. These features are
then assigned importance using random forest classifer. The measure of importance used in this implementation
is mean decrease in accuracy.
}
\examples{
data(pitlatrine)
physeq <- taxa_level(pitlatrine,"Phylum")
deseq_sig  <- differential_abundance(physeq, grouping_column = "Country")
#plot the significant features
plot_signif(deseq_sig$plotdata) #see function \\link[microbiomeSeq]{plot_signif}

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},  Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
