% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDA.R
\name{KDA}
\alias{KDA}
\title{kernel-based differential analysis}
\usage{
KDA(physeq, grouping_column, analyse = "abundance", method = "dscore",
  p.adjust.method = "BH", corr = 0.99,
  adjusted.p.value.threshold = 0.05, select.variables = NULL, ...)
}
\arguments{
\item{physeq}{(Required). A \link[phyloseq]{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the.
information, this should be one of the components of grouping vector.}

\item{analyse}{(optional). A character string specifying whether to analyse taxa abundance ("abundance") or sample data ("meta"). Default is set to analyse taxa abundance.}

\item{method}{(optional). A character string specifying the kernel based method to be used for differential analysis,
two options are available for this; "dscore" and "sscore". See \link[KMDA]{dscore} and  \link[KMDA]{sscore} for details.}

\item{p.adjust.method}{(optional). method to adjust raw pvalues obtained by the distance based score tests.}

\item{corr}{(optional). Threshold correlation on which grouping of features/ variables is based.}

\item{adjusted.p.value.threshold}{Cut off p-value for significance of differentially abundant/correlated taxa/environmental variables, default is 0.05.}
}
\value{
Returns a list of two items: \itemize{
        \item  kscore_table: A \code{data.frame} of taxa/ environmental variable with kernel based score stats, raw and adjusted pvalues.
        \item plotdata: A \code{data.frame} similar to kcore_table but organised in a way compatible to ggplot.
       }
}
\description{
This function performs differential analysis using a distance-based kernel score test.
It obtains set(s) of differentially abundant features between a specified pair of conditions/groups.
The sets are obtained by grouping based on correlation of abundance among features. The strength of the relationship that must exist for
features/variables to belong to the same set can be specified. A similar approach applies to environmental
variables which can also be grouped into sets depending on a desired level of correlation amongst member variables.
}
\examples{
data(pitlatrine)
physeq<-data(pitlatrine)
physeq<- taxa_level(physeq, "Phylum")
kda_sig  <- KDA(physeq,grouping_column = "Country")
plot_kda(kda_sig$plotdata)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},  Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
