% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_abundance.R
\name{kruskal_abundance}
\alias{kruskal_abundance}
\title{Kruskal-Wallis differential abundance  analysis}
\usage{
kruskal_abundance(physeq, grouping_column, pvalue.threshold = 0.05)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the information.
information, this should be one of the components of grouping vector.}

\item{pvalue.threshold.}{Cut off p-value for significance of differentially abundant taxa, default is 0.05.}
}
\value{
Returns a list of three items: \itemize{
        \item  SignfeaturesTable: A \code{data.frame} of taxa with coresponding raw p-values, corrected p-values, family wise error rate and expected abundance
        computed by using raw p-values.
        \item  importance: A \code{data.frame} of taxa mean decrease in accuracy as obtained by random forest classifier.
        \item plotdata: A \code{data.frame} of taxa and corresponding corrected p-values, importance rank organised
        in form accepted by ggplot.
       }
}
\description{
This function finds the features that are significantly
differentially abundant  in the provided taxa abundance data under different conditions
using Kruskal-Wallis test. The p-values values generated are corrected for multiple testing
using  family wise error rate. Significance is based on the corrected pvalue threshold. Significant features
are assigned importance using random forest classifier. The measure of importance used in this case is mean decrese accuracy.
}
\examples{
data(pitlatrine)
physeq <- pitlatrine
physeq <- taxa_level(physeq,"Phylum")
kw_sig <-  kruskal_abundance(physeq, "Country")
#plot the significant features
plot_signif(kw_sig$plotdata) #see function \\link[microbiomeSeq]{plot_signif}

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015

\url{http://www.bigre.ulb.ac.be/courses/statistics_bioinformatics/practicals/microarrays_berry_2010/berry_feature_selection.html}
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
