% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_taxa_cor.R
\name{module_env_correlation}
\alias{module_env_correlation}
\title{Relationship between sub-communities and environment}
\usage{
module_env_correlation(co_occur_res, select.variables = NULL, ...)
}
\arguments{
\item{co_occur_res}{an object returned by `co_occurence_network` function.}

\item{select.variables}{environmental variables of interest.}
}
\description{
This function calculates correlation between modules (sub-communities) and environmental
variables. The most connected node(taxon) in a given module is considered a good representation
of the module. This is implemented in accordance to (Deng et al. 2012) where
correlations between module-based eigengenes and environmental factors are
used to detect the modules' response to environmental change.
}
\examples{
mod.env.cor <- module_env_correlation(co_occur)
p <- plot_taxa_env(mod.env.cor)
print(p)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015

Deng, Ye, Yi-Huei Jiang, Yunfeng Yang, Zhili He, Feng Luo, and Jizhong Zhou. 2012.
“Molecular Ecological Network Analyses.” BMC Bioinformatics 13 (1). BioMed Central: 113.
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},  Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
