% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalise_data}
\alias{normalise_data}
\title{Data normalisation}
\usage{
normalise_data(physeq, norm.method, ...)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{method}{(Required). A \code{method} to used to normalise the data. Available methods include:
"edgeRnorm", "varstab", "randomsubsample", "proportion" .}
}
\value{
Returns a phyloseq object whose abundance data is normalised.
}
\description{
This function normalises taxa abundance
}
\examples{
data(pitlatrine)
physeq<-data(pitlatrine)
physeq <- normalise_data(physeq,method = "randomsubsample")
physeq <- normalise_data(physeq,method = "edgeRnorm")
physeq <- normalise_data(physeq,method = "proportion")

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
