% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_anova.R
\name{perform_anova}
\alias{perform_anova}
\title{pairwise ANOVA}
\usage{
perform_anova(df, meta_table, grouping_column, pValueCutoff)
}
\arguments{
\item{df}{(Required). A \code{data.frame} containg measures a measure to be analysed.}

\item{meta_table}{(Required). A data frame containing atleast one variable (grouping variable).}

\item{grouping_column}{(Required). A character string specifying name of the grouping variable in the supplied meta table.}
}
\value{
It returns a list of two data frames: One being p-values obtained from the pairwise ANOVA of
the measure and levels of grouping variable , the other is containing updated measure.
}
\description{
This function performs ANOVA of a given measure in specified groups, in addition it
also performs pairwise ANOVA of the measure between possible pairs of levels in the grouping variable. It returns
p-values obtained.
}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
