% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_diversity.R
\name{plot_anova_diversity}
\alias{plot_anova_diversity}
\title{Alpha diversity with ANOVA}
\usage{
plot_anova_diversity(physeq, method, grouping_column,
  pValueCutoff = 0.05, outfile = "anova_diversity.csv")
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{method}{(Required). A list of character strings specifying \code{method} to be used to calculate for alpha diversity
in the data. Available methods are: "richness", "fisher", "simpson", "shannon" and "evenness".}

\item{grouping_column}{(Required). A character string specifying the name of a categorical variable containing  grouping information.}
}
\value{
Returns a ggplot object which can further be manipulated further.
}
\description{
This function calculates alpha diversity of provided community data using
selected indices/method(s). It performs pair-wise ANOVA of diversity measures between groups
and outputs a plot for each of the selected methods(indices) annotated with significance levels.
}
\examples{
data(pitlatrine)
physeq <- pitlatrine
p<-plot_anova_diversity(physeq, method = c("richness","simpson"),grouping_column =  "Country",pValueCutoff=0.05)
plot_anova_diversity(physeq, method = c("richness","shannon"), grouping_column = "Depth")
print(p)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
