% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_env.R
\name{plot_anova_env}
\alias{plot_anova_env}
\title{ANOVA of environmental variables}
\usage{
plot_anova_env(physeq, grouping_column, pValueCutoff = 0.05,
  select.variables = NULL)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{grouping_column}{(Required). Name of a categorical variable that is preffered for grouping the.
information, this should be one of the components of grouping vector.}

\item{pValueCutoff.}{p-value threshold for significance of anova results, default set to 0.05.}

\item{select.variables.}{A vector of character strings(s) specifying environmental variable(s) to be analysed. If
not supplied, all numeric variables are analysed.}
}
\value{
Returns a ggplot object. This can further be manipulated as preferred by user.
}
\description{
This functions applies analysis of variance on the measured environmental variables
in specified groups and constructs a plot showing how the environmental variables vary in the groups.
The plot is annotated with significance level as obtained from \code{anova}.
}
\examples{
data(pitlatrine)
physeq<-pitlatrine
p1<-plot_anova_env(physeq,grouping_column =  "Country",select.variables=c("Temp","pH"))
print(p1)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
