% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cca.R
\name{plot_cca}
\alias{plot_cca}
\title{Canonical Correspondence Analysis plot}
\usage{
plot_cca(physeq, grouping_column, pvalueCutoff = 0.01,
  norm_method = NULL, env.variables = NULL, num.env.variables = NULL,
  exclude.variables = NULL, draw_species = F)
}
\arguments{
\item{`physeq`}{is a required phyloseq object containing taxa abundance and meta data.}

\item{`grouping_column`}{is the variable in the meta data with respect to which the data should be grouped,}

\item{`pvalueCutoff`}{the threshold p-value in `anova` of distance matrices, default set to `0.05`.}

\item{`env.variables`}{is a list of variables prefered to be on the cca plot.}

\item{`exclude.variables`}{a list of variables to be excluded from the cca plot.}

\item{`num.env.variables`}{is an integer specifying the number of variables to show on the cca plot.}
}
\value{
A ggplot object
}
\description{
This function finds a set of best environmental variables that describe community structure
superimposes them on the ordination plot.
}
\examples{
data(pitlatrine)
physeq<-data(physeq)
ccaplot  <- plot_cca(physeq, grouping_column = "Country")

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},  Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
