% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signif.R
\name{plot_signif}
\alias{plot_signif}
\title{Plotting Significant features data}
\usage{
plot_signif(df = NULL, top.taxa = 20, ...)
}
\arguments{
\item{df}{(optional). A \code{data.frame} returned by \link[microbialSeq]{kruskal_expression},
\link[microbialSeq]{differential_abundance}}

\item{top.taxa}{An integer specifying the number of top taxa to be visualised.}

\item{df_accuracy(optional).}{A \code{data.frame} of  feature importance information.}

\item{res_tax}{(optional). A \code{data.frame} bearing results for plotting MA plot.}

\item{corrections_table}{(optional). A \code{data.frame} containing multiple testing corrections information.}
}
\value{
Returns ggplot object(s). This can further be manipulated as preferred by user.
}
\description{
This function takes data generated from differential expression analysis functions (see \link[microbialSeq]{kruskal_expression},
\link[microbialSeq]{differential_abundance}) and returns ggplot objects showing different visualisation given the available parameters.
}
\examples{
p <- plot_signif(deseq_sig$plotdata)
print(p)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
