% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa}
\alias{plot_taxa}
\title{Local Contribution to Beta Diversity (LCBD)}
\usage{
plot_taxa(physeq, grouping_column, method = "hellinger",
  number.taxa = 21, filename = NULL)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{method.}{A \code{method} to used to calculate dissimilarity coefficents.
Available methods are: "hellinger", "chord", "chisquare", "profiles",
"percentdiff", "ruzicka", "divergence", "canberra",
"whittaker", "wishart", "kulczynski", "jaccard", "sorensen", "ochiai",
"ab.jaccard", "ab.sorensen", "ab.ochiai", "ab.simpson", "euclidean". the function uses
"hellinger" as a default.}

\item{grouping_column(Required).}{Name of a categorical variable that is preffered for grouping the.
information.}
}
\value{
Returns a ggplot object. This can further be manipulated as preferred by user.
}
\description{
This function finds the most abundant taxa  for each sample. It also calculates local
contribution to beta  diversity using a selected dissimilarity coefficient.
It returns a ggplot object which a visual representation of the most abundant
taxa for each of the samples. The number of top taxa can be suggested as an argument. The visual
representation is limited to 21 top taxa, more information can be availed by supplying a file name to which
details of LCBD can be written.
}
\examples{
plot_taxa(physeq, "Country")

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
