% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest_res.R
\name{randomforest_res}
\alias{randomforest_res}
\title{Importance of significant features.}
\usage{
randomforest_res(data, groups)
}
\arguments{
\item{data}{(Required). A \code{data.frame} of features to be assigned importance.}

\item{groups}{(Required). Vector corresponding to groups/conditions or source of variation in data.}
}
\value{
Returns a \code{data.frame} of importance measure (mean decrease in accuracy) and ranks.
}
\description{
This function uses Breiman's random forest algorithm that can be used in unsupervised mode for assessing proximities among data points.
Random forest classifier is used to determine the importance of differentially expressed features/taxa to the microbial community.
The measure used in this case is Mean Descrease in Accuracy aso know as permutation importance which is
reported for each of the features. This is obtained by removing the relationship of a feature  and measuring increase in error.
Consequently, the feature with high mean decrease in accuracy is considered most important.
}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com},  Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
