% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_taxa_correlation.R
\name{taxa.env.correlation}
\alias{taxa.env.correlation}
\title{Taxa-Environment correlation}
\usage{
taxa.env.correlation(physeq, grouping_column, method = "pearson",
  pvalue.threshold = 0.05, padjust.method = "BH", adjustment = 1,
  num.taxa = 50, select.variables = NULL)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the information.
information, this should be one of the components of grouping vector.}

\item{method}{A character string indicating which correlation coefficient is to be computed, available options are "pearson" which is also the default, "kendall" and "spearman".}

\item{adjustment}{(optional). An integer to specify how the p-values should be adjusted;
\itemize{
\item 1 - donot adjust
\item 2 - adjust environmental variables + Groups (column on the correlation plot)
\item 3 - adjust Taxa + Groups (row on the correlation plot for each Groups)
\item 4 - adjust Taxa (row on the correlation plot)
\item 5 - adjust environmental variables (panel on the correlation plot)
}}

\item{select.variables}{(optional). Character string for the names(s) of environmental variables to be considered in correlation compution.
if not given, all numerical variables are considered.}

\item{pvalue.threshold.}{Cut off p-value for significance of correlation between taxa abundance and environmental variables, default is 0.05.}

\item{padjust.method.}{Method for adjusting p-values.}
}
\value{
A \code{data.frame} of correlation results.
}
\description{
This function finds the relationship between most abundant taxa
and numerical environmental variables based on correlation. The abundance of each feature/taxa is
correlated with each of the environmental variables. A correlation test is performed and associated
p-values are adjusted for multiple testing. The scheme of adjustment is elaborated in the arguments section. The function returns a data.frame with raw p-values, corrected p-values,
and correlation results.
}
\examples{
data(pitlatrine)
physeq <- pitlatrine
physeq <- taxa_level(physeq, "Phylum")
tax_env_cor <- taxa.env.correlation(physeq,"Country")
plot_taxa_env(tax_env_cor)

}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Alfred Ssekagiri \email{assekagiri@gmail.com}, Umer Zeeshan Ijaz \email{Umer.Ijaz@glasgow.ac.uk}
}
