% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_level.R
\name{taxa_level}
\alias{taxa_level}
\title{Taxa level collating}
\usage{
taxa_level(physeq, which_level)
}
\arguments{
\item{physeq}{(Required). A \link[phyloseq]{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{which_level}{(Required). Character string specifying taxonomic level.}
}
\value{
physeq object at specified taxonomic level.
}
\description{
This function takes a physeq object and returns a physeq object with taxa at a specified taxonomic level.
It extends the \link[phyloseq]{tax_glom}  to include names of corresponding taxa
at a specified taxonomy level and  updating tree tip labels accordingly.
}
\examples{

data(pitlatrine)
physeq<-data(physeq)
physeq <- taxa_level(physeq = physeq,which_level = "Family")

}
